Add CheckError extension method for Glfw to GlfwExtensions library
